(MAINT) update filter spec names
